Update README.md
Fix the example x coordinates in the README.md. The x coordinates that where before generates the dmg with the "Applications" folder pointing to the "App". It should be backwards, the "App" pointing to the "Applications" folder.

Update README.md

Remove empty line after ``` // `type: link` creates a link to the specified target```