Use Node's `fs.copyFile`, if available.
`cp-file` doesn't seem to work on my system (macOS 10.14.4, Node 11.5.0 from Homebrew). It either throws an error (and trying to get the `stack` from that error results in *another* error being thrown!), or silently fails.

Node's `fs.copyFile` has no such problem. It's not available before Node 8.5.0, and appdmg targets Node ≥ 4, so this commit changes the code so that `fs.copyFile` is used if it is available, and if not, `cp-file` is used as a fallback.